[svn-r3079] Purpose: bug fix - by C API
Description:
	The prototype of two C API functions, H5Tget_offset and
	H5Sget_simple_extent_npoints was changed to fix bug #446,
	resulting in the need for updating the two corresponding
	C++ API functions.

Solution:
	- Changed the return type of AtomType::getOffset from size_t
	  to int, and the error return value from 0 to -1.
	- Changed the return type of DataSpace::getSimpleExtentNpoints
	  from hsize_t to hssize_t, and the error value from 0 to -1.

Platforms tested:
	arabica (sparc-sun-solaris 2.7)

+++ c++/src/H5AtomType.h
@@ -25,7 +25,8 @@ class AtomType : public DataType {
 	void setPrecision( size_t precision ) const;
 
 	// Retrieves the bit offset of the first significant bit. 
-	size_t getOffset() const;
+	// 12/05/00 - changed return type to int from size_t - C API
+	int getOffset() const;
 
 	// Sets the bit offset of the first significant bit. 
 	void setOffset( size_t offset ) const;
